16-may-24-2

+++ op/lab1.3/test.cpp
@@ -2,7 +2,7 @@
 #include <fstream>
 #include <algorithm>
 
-int *exclude_max(const int *A, int size, int &new_size) {
+int *exclude_max(const int *A, size_t size, size_t &new_size) {
     // Находим максимальный элемент в массиве A
     int max_element = *std::max_element(A, A + size);
 
@@ -11,7 +11,7 @@ int *exclude_max(const int *A, int size, int &new_size) {
     new_size = 0;
 
     // Проходим по массиву A и добавляем элементы в B, исключая максимальный элемент
-    for (int i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         if (A[i] != max_element)
             B[new_size++] = A[i];
     }
@@ -28,23 +28,23 @@ int main() {
         return 1;
     }
 
-    int n;
+    size_t n;
     input_file >> n; // Читаем размерность массива
 
     int *A = new int[n];
 
     // Считываем элементы массива A из файла
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
         input_file >> A[i];
     }
 
-    int new_size;
+    size_t new_size;
     // Получаем выходной массив, исключая максимальный элемент
     int *B = exclude_max(A, n, new_size);
 
     // Записываем размерность и элементы выходного массива в файл
     output_file << new_size << std::endl;
-    for (int i = 0; i < new_size; ++i) {
+    for (size_t i = 0; i < new_size; ++i) {
         output_file << B[i] << " ";
     }
